Move unnecessary templates in headless package to main package
