+++ src/converse-register.js
@@ -12,12 +12,12 @@
 (function (root, factory) {
     define(["utils/form",
             "@converse/headless/converse-core",
-            "@converse/headless/templates/form_username.html",
+            "templates/form_username.html",
             "templates/register_link.html",
             "templates/register_panel.html",
             "templates/registration_form.html",
             "templates/registration_request.html",
-            "@converse/headless/templates/form_input.html",
+            "templates/form_input.html",
             "templates/spinner.html",
             "converse-controlbox"
     ], factory);
